Refactor OandaBrokerage to support different API versions
