+++ Brokerages/Oanda/RestV1/DataType/Candle.cs
@@ -17,7 +17,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-namespace QuantConnect.Brokerages.Oanda.DataType
+namespace QuantConnect.Brokerages.Oanda.RestV1.DataType
 {
 #pragma warning disable 1591
     public struct Candle
