+++ Brokerages/Oanda/RestV1/DataType/Communications/OrdersResponse.cs
@@ -16,9 +16,10 @@
  * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, 
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
+
 using System.Collections.Generic;
 
-namespace QuantConnect.Brokerages.Oanda.DataType.Communications
+namespace QuantConnect.Brokerages.Oanda.RestV1.DataType.Communications
 {
 #pragma warning disable 1591
     /// <summary>
