+++ Brokerages/Oanda/RestV1/DataType/Communications/Response.cs
@@ -16,11 +16,12 @@
  * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, 
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
+
 using System.Reflection;
 using System.Text;
-using QuantConnect.Brokerages.Oanda.Framework;
+using QuantConnect.Brokerages.Oanda.RestV1.Framework;
 
-namespace QuantConnect.Brokerages.Oanda.DataType.Communications
+namespace QuantConnect.Brokerages.Oanda.RestV1.DataType.Communications
 {
 #pragma warning disable 1591
     /// <summary>
