+++ Brokerages/Oanda/RestV1/DataType/TradeData.cs
@@ -16,9 +16,10 @@
  * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, 
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
-using QuantConnect.Brokerages.Oanda.DataType.Communications;
 
-namespace QuantConnect.Brokerages.Oanda.DataType
+using QuantConnect.Brokerages.Oanda.RestV1.DataType.Communications;
+
+namespace QuantConnect.Brokerages.Oanda.RestV1.DataType
 {
 #pragma warning disable 1591
     /// <summary>
