+++ Brokerages/Oanda/RestV1/Framework/Common.cs
@@ -16,10 +16,11 @@
  * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, 
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
+
 using System;
 using System.Reflection;
 
-namespace QuantConnect.Brokerages.Oanda.Framework
+namespace QuantConnect.Brokerages.Oanda.RestV1.Framework
 {
 #pragma warning disable 1591
     /// <summary>
